refactor(desktop): split frame codec, sidecar IPC, and event sink out of the god-files
Extract three self-contained seams identified in #282, #295, #296:

- desktop/rust: frame.rs -- the sidecar wire codec (consts, encode/decode,
  varint state machine, sync + async read/write twins) plus its five dedicated
  tests. The async twins keep their #[cfg(any(windows, test))] gate so drift
  still fails CI off-Windows. (#282)

- desktop/rust: sidecar_ipc/{mod,unix,windows}.rs -- the cfg-gated sidecar-IPC
  twins (connect_sidecar_endpoint, is_sidecar_gone, endpoint_holder_pid, the
  dev-endpoint helpers, the peer-credential checks) grouped as a module
  directory so the two OS implementations are diffable side-by-side. The
  Windows half also absorbs the named-pipe transport primitives
  (pipe_runtime, SyncPipeReader/Writer, PipeConnect, SID helpers) it sits on.
  (#296)

- desktop/go: event_sink.go -- App's event pub-sub (generic sink, relay-aware
  sink, and emitForOwner) lifted into its own type with its own mutex, so
  App's mutex surface drops to the lifecycle/config locks. App keeps thin
  forwarders so every existing call site is unchanged. (#295)

refactor(desktop): close Windows handshake SID gap and dedup GetSidecarInfo

Deep-review follow-through on the #282/#295/#296 extraction branch.

Security + leak (the substantive change):
- windows_handshake_async called open_named_pipe_client directly, bypassing
  connect_sidecar_endpoint -- so the Windows dev-sidecar handshake enforced
  NO same-user SID check (Unix enforced it via connect_sidecar_endpoint ->
  require_same_user_peer). Add connect_sidecar_endpoint_async, the async
  twin that returns the raw NamedPipeClient the handshake needs while running
  require_same_user_pipe_peer, so the handshake adopts no unverified peer.
- Route windows_handshake_async through it. This also closes the
  transport-layer leak (FINDING-B/DESIGN-1) and makes the busy-pipe error
  literal single-source.
- Extract get_sidecar_info_request into frame.rs and use it at all 5
  GetSidecarInfo sites (FINDING-D/DESIGN-3).

Doc/import fixes recovered from the review pass:
- Restore the dropped rationale docs: private_dev_sidecar_endpoint (unix.rs),
  the FILE_OBJECT-lock/deadlock architectural banner (windows.rs module doc),
  and fix the dangling "see there" cross-ref.
- Fix stale comments: the alloc_probe doc now points at frame::tests (the test
  moved in the extraction); a misplaced import comment; and the
  MAX_VARINT_BYTES comment alignment.
- Fix a Windows compile regression the extraction introduced: the move deleted
  the NamedPipeClient/ClientOptions imports but windows_handshake_async and a
  windows test still referenced them.

The Windows handshake/SID change is not compilable or testable on this macOS
host; it is grounded in the verified caller map (require_same_user_pipe_peer
borrows the client immutably, does no I/O, and the client stays usable for
handshake I/O afterward -- the same pattern connect_sidecar_endpoint already
uses) and matches Unix's existing behavior. Windows CI is the validation gate.

Verified locally: cargo +1.97.1 clippy --all-targets -- -D warnings clean;
55 tests pass (was 54; +1 for get_sidecar_info_request); cargo fmt clean;
go vet + go test green.
